Extract server settings module from main.rs
