+++ src/main.rs
@@ -22,7 +22,17 @@ use serde_json::json;
 use sha2::{Digest, Sha256};
 
 use crate::auth::{AuthConfig, LoginRequest};
+#[cfg(test)]
 use crate::builtin_detection::JcodeDetectionVariant;
+use server_settings::{
+    apply_cli_overrides, load_runtime_server_settings, log_event, save_runtime_server_settings,
+    server_settings_path, settings_public_json, BackendMode, LogLevel, RecentWorkspace,
+    RuntimeServerSettings,
+};
+#[cfg(test)]
+use server_settings::{
+    default_runtime_server_settings, validate_runtime_server_settings, PersistedServerSettings,
+};
 
 mod assets;
 mod auth;
@@ -34,6 +44,7 @@ mod file_browser;
 mod git_ui;
 mod lsp;
 mod protocol;
+mod server_settings;
 mod service;
 mod terminal_text;
 
@@ -162,40 +173,6 @@ struct WebConfig {
     tls: TlsConfig,
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "kebab-case")]
-enum BackendMode {
-    ExternalHerdr,
-    Builtin,
-    Auto,
-}
-
-impl BackendMode {
-    fn parse(value: &str) -> io::Result<Self> {
-        match value {
-            "external-herdr" | "external" | "herdr" => Ok(Self::ExternalHerdr),
-            "builtin" | "built-in" => Ok(Self::Builtin),
-            "auto" => Ok(Self::Auto),
-            other => Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("invalid --backend-mode: {other}; use external-herdr, builtin, or auto"),
-            )),
-        }
-    }
-
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::ExternalHerdr => "external-herdr",
-            Self::Builtin => "builtin",
-            Self::Auto => "auto",
-        }
-    }
-
-    fn is_builtin(self) -> bool {
-        matches!(self, Self::Builtin)
-    }
-}
-
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 enum SessionBackendTarget {
     ExternalHerdr,
@@ -226,92 +203,12 @@ struct TlsConfig {
     key_path: Option<PathBuf>,
 }
 
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-struct RecentWorkspace {
-    path: String,
-    label: Option<String>,
-    branch: Option<String>,
-    kind: Option<String>,
-    opened_at: Option<u64>,
-}
-
 #[derive(Clone, Debug, PartialEq, Eq)]
 enum TlsMode {
     Off,
     Auto,
     SelfSigned,
     Files,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-struct PersistedServerSettings {
-    bind: Option<String>,
-    user: Option<String>,
-    password: Option<String>,
-    localhost_no_auth: Option<bool>,
-    no_sleep_auto_cooldown_seconds: Option<u64>,
-    backend_mode: Option<BackendMode>,
-    builtin_shell: Option<String>,
-    default_folder: Option<String>,
-    builtin_backend_enabled: Option<bool>,
-    external_herdr_backend_enabled: Option<bool>,
-    jcode_detection_variant: Option<String>,
-    log_level: Option<LogLevel>,
-    #[serde(default)]
-    lsp: Option<lsp::LspSettings>,
-    #[serde(default)]
-    recent_workspaces: Option<Vec<RecentWorkspace>>,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-enum LogLevel {
-    #[default]
-    None,
-    Info,
-    Debug,
-}
-
-impl LogLevel {
-    fn as_str(&self) -> &'static str {
-        match self {
-            LogLevel::None => "none",
-            LogLevel::Info => "info",
-            LogLevel::Debug => "debug",
-        }
-    }
-
-    fn enabled(&self) -> bool {
-        !matches!(self, LogLevel::None)
-    }
-}
-
-fn log_event(level: &LogLevel, message: &str) {
-    if level.enabled() {
-        let secs = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs())
-            .unwrap_or(0);
-        eprintln!("[{secs}] {message}");
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-struct RuntimeServerSettings {
-    bind: SocketAddr,
-    user: Option<String>,
-    password: Option<String>,
-    localhost_no_auth: bool,
-    no_sleep_auto_cooldown_seconds: u64,
-    backend_mode: BackendMode,
-    builtin_shell: Option<String>,
-    default_folder: String,
-    builtin_backend_enabled: bool,
-    external_herdr_backend_enabled: bool,
-    jcode_detection_variant: JcodeDetectionVariant,
-    log_level: LogLevel,
-    lsp: lsp::LspSettings,
-    recent_workspaces: Vec<RecentWorkspace>,
 }
 
 struct NoSleepGuard {
@@ -679,68 +576,6 @@ impl EventStream {
     }
 }
 
-impl AuthConfig {
-    fn from_settings(settings: &RuntimeServerSettings) -> io::Result<Self> {
-        validate_runtime_server_settings(settings)?;
-        Ok(crate::auth::AuthConfig::from_parts(
-            settings.user.clone(),
-            settings.password.clone(),
-            settings.localhost_no_auth,
-        ))
-    }
-}
-
-fn validate_runtime_server_settings(settings: &RuntimeServerSettings) -> io::Result<()> {
-    let local_bind = settings.bind.ip().is_loopback();
-    if !local_bind && (settings.user.is_none() || settings.password.is_none()) {
-        return Err(io::Error::new(
-            io::ErrorKind::PermissionDenied,
-            "username and password are required before binding to 0.0.0.0 or any non-local address",
-        ));
-    }
-    if local_bind
-        && !settings.localhost_no_auth
-        && (settings.user.is_none() || settings.password.is_none())
-    {
-        return Err(io::Error::new(
-            io::ErrorKind::PermissionDenied,
-            "set username/password or allow localhost auth bypass",
-        ));
-    }
-    if settings.no_sleep_auto_cooldown_seconds > 3600 {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "no-sleep auto cooldown must be 3600 seconds or less",
-        ));
-    }
-    if !settings.builtin_backend_enabled && !settings.external_herdr_backend_enabled {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "at least one backend type must be enabled",
-        ));
-    }
-    Ok(())
-}
-
-fn default_runtime_server_settings(bind: SocketAddr) -> RuntimeServerSettings {
-    RuntimeServerSettings {
-        bind,
-        user: None,
-        password: None,
-        localhost_no_auth: true,
-        no_sleep_auto_cooldown_seconds: 60,
-        backend_mode: BackendMode::Builtin,
-        builtin_shell: None,
-        default_folder: default_working_folder(None),
-        builtin_backend_enabled: true,
-        external_herdr_backend_enabled: true,
-        jcode_detection_variant: JcodeDetectionVariant::default(),
-        log_level: LogLevel::default(),
-        lsp: lsp::LspSettings::default(),
-        recent_workspaces: Vec::new(),
-    }
-}
-
 fn readable_directory(path: &Path) -> io::Result<PathBuf> {
     let metadata = fs::metadata(path)?;
     if !metadata.is_dir() {
@@ -835,168 +670,6 @@ fn request_default_folder_permission(default_folder: &Path) -> io::Result<PathBu
         io::ErrorKind::PermissionDenied,
         format!("cannot read {}", default_folder.display()),
     ))
-}
-
-fn server_settings_path() -> PathBuf {
-    if let Ok(dir) = std::env::var("XDG_CONFIG_HOME") {
-        return PathBuf::from(dir).join("herdr-webui/webui-settings.json");
-    }
-    std::env::var("HOME")
-        .map(|home| PathBuf::from(home).join(".config/herdr-webui/webui-settings.json"))
-        .unwrap_or_else(|_| std::env::temp_dir().join("herdr-webui/webui-settings.json"))
-}
-
-/// Apply CLI flags that must win over persisted settings.
-/// An explicit `--bind` beats the persisted bind from webui-settings.json;
-/// otherwise a preview instance could silently squat the saved port instead of
-/// the one the operator asked for.
-fn apply_cli_overrides(settings: &mut RuntimeServerSettings, config: &WebConfig) {
-    if config.bind_explicit {
-        settings.bind = config.bind;
-    }
-}
-
-fn load_runtime_server_settings(default_bind: SocketAddr) -> io::Result<RuntimeServerSettings> {
-    let mut settings = default_runtime_server_settings(default_bind);
-    let path = server_settings_path();
-    let Ok(raw) = fs::read_to_string(path) else {
-        save_runtime_server_settings(&settings)?;
-        return Ok(settings);
-    };
-    let raw_json: serde_json::Value = serde_json::from_str(&raw).map_err(|err| {
-        io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("invalid webui-settings.json: {err}"),
-        )
-    })?;
-    let missing_keys = [
-        "bind",
-        "user",
-        "password",
-        "localhost_no_auth",
-        "no_sleep_auto_cooldown_seconds",
-        "backend_mode",
-        "builtin_shell",
-        "default_folder",
-        "builtin_backend_enabled",
-        "external_herdr_backend_enabled",
-        "log_level",
-    ]
-    .iter()
-    .any(|key| raw_json.get(key).is_none());
-    let persisted: PersistedServerSettings = serde_json::from_value(raw_json).map_err(|err| {
-        io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("invalid webui-settings.json: {err}"),
-        )
-    })?;
-    if let Some(bind) = persisted.bind {
-        settings.bind = bind.parse().map_err(|err| {
-            io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("invalid saved bind: {err}"),
-            )
-        })?;
-    }
-    if persisted.user.is_some() {
-        settings.user = persisted.user.filter(|value| !value.is_empty());
-    }
-    if persisted.password.is_some() {
-        settings.password = persisted.password.filter(|value| !value.is_empty());
-    }
-    if let Some(localhost_no_auth) = persisted.localhost_no_auth {
-        settings.localhost_no_auth = localhost_no_auth;
-    }
-    if let Some(cooldown) = persisted.no_sleep_auto_cooldown_seconds {
-        settings.no_sleep_auto_cooldown_seconds = cooldown;
-    }
-    if let Some(backend_mode) = persisted.backend_mode {
-        settings.backend_mode = backend_mode;
-    }
-    if persisted.builtin_shell.is_some() {
-        settings.builtin_shell = persisted.builtin_shell.filter(|value| !value.is_empty());
-    }
-    if persisted.default_folder.is_some() {
-        settings.default_folder =
-            default_working_folder_without_prompt(persisted.default_folder.as_deref());
-    }
-    if let Some(enabled) = persisted.builtin_backend_enabled {
-        settings.builtin_backend_enabled = enabled;
-    }
-    if let Some(enabled) = persisted.external_herdr_backend_enabled {
-        settings.external_herdr_backend_enabled = enabled;
-    }
-    if let Some(variant_str) = persisted.jcode_detection_variant {
-        settings.jcode_detection_variant = JcodeDetectionVariant::from_str(&variant_str);
-    }
-    if let Some(log_level) = persisted.log_level {
-        settings.log_level = log_level;
-    }
-    if let Some(lsp) = persisted.lsp {
-        settings.lsp = lsp;
-    }
-    if let Some(recent) = persisted.recent_workspaces {
-        settings.recent_workspaces = recent;
-    }
-    validate_runtime_server_settings(&settings)?;
-    if missing_keys {
-        save_runtime_server_settings(&settings)?;
-    }
-    Ok(settings)
-}
-
-/// Tests must never persist settings to the operator's real config file:
-/// a stray save clobbers `~/.config/herdr-webui/webui-settings.json` and
-/// locks the operator out of their own server (username/password fixtures
-/// replace real credentials). Every test that reaches a persisting route
-/// must set `XDG_CONFIG_HOME` to a temp dir; `lock_env()` serializes those
-/// env mutations. This guard turns an unisolated test into a loud failure
-/// instead of a silent config overwrite.
-#[cfg(test)]
-fn assert_test_settings_isolation() {
-    if std::env::var_os("XDG_CONFIG_HOME").is_none() {
-        panic!(
-            "test would write the real settings path; set XDG_CONFIG_HOME \
-             to a temp dir (see lock_env-isolated tests) before persisting"
-        );
-    }
-}
-
-#[cfg(not(test))]
-fn assert_test_settings_isolation() {}
-
-fn save_runtime_server_settings(settings: &RuntimeServerSettings) -> io::Result<()> {
-    validate_runtime_server_settings(settings)?;
-    // Runs after validation: rejection tests assert on the validation error
-    // itself and never write, so they must not trip the guard.
-    assert_test_settings_isolation();
-    let path = server_settings_path();
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)?;
-    }
-    let content = serde_json::to_string_pretty(&PersistedServerSettings {
-        bind: Some(settings.bind.to_string()),
-        user: settings.user.clone(),
-        password: settings.password.clone(),
-        localhost_no_auth: Some(settings.localhost_no_auth),
-        no_sleep_auto_cooldown_seconds: Some(settings.no_sleep_auto_cooldown_seconds),
-        backend_mode: Some(settings.backend_mode),
-        builtin_shell: settings.builtin_shell.clone(),
-        default_folder: Some(settings.default_folder.clone()),
-        builtin_backend_enabled: Some(settings.builtin_backend_enabled),
-        external_herdr_backend_enabled: Some(settings.external_herdr_backend_enabled),
-        jcode_detection_variant: Some(settings.jcode_detection_variant.as_str().to_string()),
-        log_level: Some(settings.log_level.clone()),
-        lsp: Some(settings.lsp.clone()),
-        recent_workspaces: Some(settings.recent_workspaces.clone()),
-    })?;
-    fs::write(&path, content)?;
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        let _ = fs::set_permissions(&path, fs::Permissions::from_mode(0o600));
-    }
-    Ok(())
 }
 
 #[tokio::main]
@@ -2019,27 +1692,6 @@ struct UpdateServerSettingsRequest {
     builtin_backend_enabled: Option<bool>,
     external_herdr_backend_enabled: Option<bool>,
     log_level: Option<LogLevel>,
-}
-
-fn settings_public_json(settings: &RuntimeServerSettings) -> serde_json::Value {
-    json!({
-        "bind": settings.bind.to_string(),
-        "username": settings.user.clone().unwrap_or_default(),
-        "has_password": settings.password.is_some(),
-        "localhost_no_auth": settings.localhost_no_auth,
-        "no_sleep_auto_cooldown_seconds": settings.no_sleep_auto_cooldown_seconds,
-        "backend_mode": settings.backend_mode.as_str(),
-        "builtin_shell": settings.builtin_shell.clone(),
-        "default_folder": settings.default_folder.clone(),
-        "builtin_backend_enabled": settings.builtin_backend_enabled,
-        "external_herdr_backend_enabled": settings.external_herdr_backend_enabled,
-        "log_level": settings.log_level.as_str(),
-        "enabled_backends": {
-            "builtin": settings.builtin_backend_enabled,
-            "external-herdr": settings.external_herdr_backend_enabled,
-        },
-        "settings_path": server_settings_path().to_string_lossy(),
-    })
 }
 
 fn no_sleep_public_json(state: &NoSleepState) -> serde_json::Value {
